fix(firestore): remove List and allow Query empty WHERE clause
